pyproject.toml
[project]
name = "segment-hunter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "colorlog>=6.9.0",
    "gcloud>=0.18.3",
    "geopandas>=1.1.1",
    "google>=3.0.0",
    "google-cloud>=0.34.0",
    "google-cloud-pubsub>=2.31.1",
    "google-cloud-secret-manager>=2.24.0",
    "google-cloud-storage>=3.4.0",
    "ipykernel>=6.30.1",
    "isort>=6.0.1",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-json-logger>=3.3.0",
    "requests>=2.32.5",
    "shapely>=2.1.1",
]
